+++ bin/simple_server.js
@@ -110,7 +110,7 @@ server.on("post_initialize", function () {
     var rootFolder = addressSpace.findObject("RootFolder");
     assert(rootFolder.browseName.toString() === "Root");
 
-    var myDevices = server.engine.addFolder(rootFolder.objects, {browseName: "MyDevices"});
+    var myDevices = addressSpace.addFolder(rootFolder.objects, {browseName: "MyDevices"});
 
     /**
      * variation 0:
@@ -119,7 +119,7 @@ server.on("post_initialize", function () {
      * Add a variable in folder using a raw Variant.
      * Use this variation when the variable has to be read or written by the OPCUA clients
      */
-    var variable0 = server.engine.addVariable({
+    var variable0 = addressSpace.addVariable({
         organizedBy: myDevices,
         browseName: "FanSpeed",
         nodeId: "ns=1;s=FanSpeed",
@@ -143,7 +143,7 @@ server.on("post_initialize", function () {
      * Avoid using this variation if the variable has to be made writable, as the server will call the getter
      * function prior to returning its value upon client read requests.
      */
-    server.engine.addVariable({
+    addressSpace.addVariable({
         organizedBy: myDevices,
         browseName: "PumpSpeed",
         nodeId: "ns=1;s=PumpSpeed",
@@ -161,7 +161,7 @@ server.on("post_initialize", function () {
         }
     });
 
-    server.engine.addVariable({
+    addressSpace.addVariable({
         organizedBy: myDevices,
         browseName: "SomeDate",
         nodeId: "ns=1;s=SomeDate",
@@ -192,7 +192,7 @@ server.on("post_initialize", function () {
         external_value_with_sourceTimestamp.sourceTimestamp = new Date();
     }, 1000);
 
-    server.engine.addVariable({
+    addressSpace.addVariable({
         organizedBy: myDevices,
         browseName: "Pressure",
         nodeId: "ns=1;s=Pressure",
@@ -215,7 +215,7 @@ server.on("post_initialize", function () {
      *
      */
 
-    server.engine.addVariable({
+    addressSpace.addVariable({
         organizedBy: myDevices,
         browseName: "Temperature",
         nodeId: "ns=1;s=Temperature",
@@ -261,7 +261,7 @@ server.on("post_initialize", function () {
     //------------------------------------------------------------------------------
     // Add a view
     //------------------------------------------------------------------------------
-    var view = server.engine.addView({
+    var view = addressSpace.addView({
         componentOf: rootFolder.views,
         browseName: "MyView"
     });