+++ test/nodeset_data_type_management/test_datatype_management.js
@@ -52,10 +52,10 @@ describe("ComplexType read from XML NodeSET file shall be binary Encodable", fun
     });
 
 
-    it("should create an enumeration from the  ServerState object", function (done) {
+    it("should create an enumeration from the ServerState object", function (done) {
 
         var test_value = nodeset.ServerState.NoConfiguration;
-        console.log(nodeset.ServerState);
+        //xx console.log(nodeset.ServerState);
         test_value.value.should.eql(2);
         done();